Move SomeSignKey and associated functions to Key module
